CLEVR/requirements.txt
cycler==0.10.0
h5py==2.10.0
kiwisolver==1.1.0
matplotlib==3.0.3
numpy==1.18.1
opt-einsum==3.1.0
Pillow==7.0.0
pyparsing==2.4.6
pyro-api==0.1.1
pyro-ppl==1.2.0
python-dateutil==2.8.1
PyYAML==5.3
six==1.14.0
torch==1.4.0
torchvision==0.1.9
tqdm==4.41.1

